requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
Django==4.2.2
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-elasticsearch-dsl==7.4
django-timezone-field==6.1.0
djangorestframework==3.14.0
elasticsearch==7.17.9
elasticsearch-dsl==7.4.1
exceptiongroup==1.2.0
flower==2.0.1
humanize==4.9.0
idna==3.6
iniconfig==2.0.0
kombu==5.3.5
packaging==23.2
pluggy==1.4.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
pytest==8.0.0
pytest-django==4.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
tornado==6.4
typing_extensions==4.9.0
tzdata==2023.4
urllib3==1.26.18
vine==5.1.0
wcwidth==0.2.13
